requirements.txt
# Core web & async
fastapi==0.115.12
uvicorn==0.34.0
httpx==0.28.1
anyio==4.8.0
asyncpg==0.30.0
SQLAlchemy==2.0.39

# Validation & environnement
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.1.0

# Telegram Bot
python-telegram-bot==22.0

# Utilitaires
requests==2.32.3
typing_extensions==4.12.2
colorama==0.4.6
click==8.1.8
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
urllib3==2.3.0
h11==0.14.0
httpcore==1.0.7
sniffio==1.3.1
greenlet==3.1.1
starlette==0.46.1
annotated-types==0.7.0
